Stop automatically converting strings to atoms in the Tesla level.
The keys that are returned from pipedrive are not constrained to a 
specific set defined by the API. Custom fields are denoted by a hex uuid
and there can be many of these. This ends up being a potential source of
exhaustion for atom space, and it makes it harder to look up/map the 
custom fields by their string key when it has been atomized.